pyproject.toml
[tool.poetry]
name = "fastapi-realworld-api"
version = "0.1.0"
description = "Exemplo de projeto utilizado no mundo real"
authors = ["Douglas <douglaspands@gmail.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.3"
sqlmodel = "^0.0.17"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
mypy = "^1.10.0"
httpx = "^0.27.0"
rich = "^13.7.1"
faker = "^25.0.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=server --cov-report html --cov-fail-under=85"
python_files = "test_*.py"
testpaths = "tests"
norecursedirs = ["__pycache__"]

[tool.poetry.scripts]
server = 'scripts.poetry:server'
server_debug = 'scripts.poetry:server_debug'
lint = 'scripts.poetry:lint'
test = 'scripts.poetry:test'
